pythonModules/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image Processing
Pillow==9.5.0
opencv-python==4.8.0.76

# Face Recognition
face-recognition==1.3.0
dlib>=19.22.0
numpy>=1.20.0

# Authentication (if needed)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Document Processing
langchain-community
docx2txt
pypdf

# AI/LLM Integration
groq

# Environment Variables
python-dotenv 